refactor: 💡 コンポーネントのディレクトリ・ファイル設計を見直した（第一弾）
