+++ pages/events-in-event/sosenkyo-book.tsx
@@ -2,11 +2,11 @@ import type { NextPage } from 'next'
 import Link from 'next/link'
 import { TwitterTweetEmbed, TwitterTimelineEmbed } from 'react-twitter-embed'
 
-import HumbergerNavigation from '@/components/HumbergerMenu/HumbergerNavigation'
+import HumbergerNavigation from '@/components/humberger-menu/HumbergerNavigation'
 import NavBar from '@/components/NavBar'
 import SiteFooter from '@/components/SiteFooter'
 
-import Contacts from '@/components/sosenkyo-book/contacts'
+import Contacts from '@/components/events-in-event/sosenkyo-book/Contacts'
 
 import { useLocale } from '@/hooks/useLocale'
